chore: use int64 for ID fields everywhere

+++ internal/db/db_certificate_authorities.go
@@ -9,17 +9,17 @@ import (
 )
 
 type CertificateAuthority struct {
-	CertificateAuthorityID int `db:"certificate_authority_id"`
+	CertificateAuthorityID int64 `db:"certificate_authority_id"`
 
 	Status string `db:"status"`
 
-	PrivateKeyID  int   `db:"private_key_id"`
+	PrivateKeyID  int64 `db:"private_key_id"`
 	CertificateID int64 `db:"certificate_id"`
-	CSRID         int   `db:"csr_id"`
+	CSRID         int64 `db:"csr_id"`
 }
 
 type CertificateAuthorityDenormalized struct {
-	CertificateAuthorityID int    `db:"certificate_authority_id"`
+	CertificateAuthorityID int64  `db:"certificate_authority_id"`
 	Status                 string `db:"status"`
 	PrivateKeyPEM          string `db:"private_key"`
 	CertificatePEM         string `db:"certificate"`
